Django-signals_orm-0x04/requirement.txt
django
djangorestframework
django-cors-headers
celery
rabbitmq
drf-yasg
django-environ
mysqlclient
python-dotenv
drf-nested-routers
djangorestframework-simplejwt
django-filter
